+++ src/pages/SearchPage/SearchPageStyle.jsx
